+++ dev/codehost_testing/config/BUILD.bazel
@@ -3,6 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "config",
     srcs = ["config.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/dev/codehost_scenario/config",
+    importpath = "github.com/sourcegraph/sourcegraph/dev/codehost_testing/config",
     visibility = ["//visibility:public"],
 )
