+++ dev/codehost_testing/reporter.go
@@ -1,4 +1,4 @@
-package codehost_scenario
+package codehost_testing
 
 import "fmt"
